+++ tests/unit/tree/test_endpoint_node.py
@@ -62,11 +62,11 @@ class TestEndpointNode:
             node = EndpointNode(
                 {"path": "/foo", "name": "node", "requests": []}, parent=parent
             )
-            assert node.path == f"http://foo.com/api/foo"
+            assert node.path == "http://foo.com/api/foo"
 
         def test_with_trailing_slashes(self):
             parent = EndpointNode(
-                {"path": "http://foo.com/", "name": "parent-node", "requests": [],}
+                {"path": "http://foo.com/", "name": "parent-node", "requests": []}
             )
             node = EndpointNode(
                 {"path": "/foo/", "name": "node", "requests": []}, parent=parent
@@ -75,7 +75,7 @@ class TestEndpointNode:
 
         def test_with_path_not_string(self):
             parent = EndpointNode(
-                {"path": "http://foo.com/", "name": "parent-node", "requests": [],}
+                {"path": "http://foo.com/", "name": "parent-node", "requests": []}
             )
             node = EndpointNode(
                 {"path": 2, "name": "node", "requests": []}, parent=parent
@@ -84,7 +84,7 @@ class TestEndpointNode:
 
         def test_calls_evaluate(self, mocker, mock_evaluate):
             parent = EndpointNode(
-                {"path": "http://foo.com/", "name": "parent-node", "requests": [],}
+                {"path": "http://foo.com/", "name": "parent-node", "requests": []}
             )
             node = EndpointNode(
                 {"path": "/foo/", "name": "node", "requests": []}, parent=parent
@@ -109,7 +109,7 @@ class TestEndpointNode:
             node = EndpointNode(
                 {"headers": headers, "name": "node", "requests": []},
                 parent=EndpointNode(
-                    {"headers": parent_headers, "name": "parent-node", "requests": [],}
+                    {"headers": parent_headers, "name": "parent-node", "requests": []}
                 ),
             )
             assert node.headers == {"abc": "def", "xxx": "www"}
@@ -120,7 +120,7 @@ class TestEndpointNode:
             node = EndpointNode(
                 {"headers": headers, "name": "node", "requests": []},
                 parent=EndpointNode(
-                    {"headers": parent_headers, "name": "parent-node", "requests": [],}
+                    {"headers": parent_headers, "name": "parent-node", "requests": []}
                 ),
             )
             assert node.headers == {"abc": "def", "xxx": "www"}
@@ -140,7 +140,7 @@ class TestEndpointNode:
             node = EndpointNode(
                 {"params": params, "name": "node", "requests": []},
                 parent=EndpointNode(
-                    {"params": parent_params, "name": "parent-node", "requests": [],}
+                    {"params": parent_params, "name": "parent-node", "requests": []}
                 ),
             )
             assert node.params == {"abc": "def", "xxx": "www"}
@@ -151,7 +151,7 @@ class TestEndpointNode:
             node = EndpointNode(
                 {"params": params, "name": "node", "requests": []},
                 parent=EndpointNode(
-                    {"params": parent_params, "name": "parent-node", "requests": [],}
+                    {"params": parent_params, "name": "parent-node", "requests": []}
                 ),
             )
             assert node.params == {"abc": "def", "xxx": "www"}
@@ -219,8 +219,8 @@ class TestEndpointNode:
                         {
                             "name": "foo",
                             "requests": [
-                                {"name": "First", "path": "http://foo.com/first",},
-                                {"name": "Second", "path": "http://foo.com/second",},
+                                {"name": "First", "path": "http://foo.com/first"},
+                                {"name": "Second", "path": "http://foo.com/second"},
                             ],
                         }
                     ],